requirements.txt
openai-whisper
opencc-python-reimplemented
